Add a value for the agent --launcher_dir flag.
This points the --launcher_dir to a consistent location (symlink).

Currently, all command tasks will become "over-committed" by 0.1 CPUs
and 32 MB of memory, per task, after an upgrade.  This is because the
agent generates the ExecutorInfo for command tasks with this slight
over-committment of resources.

The agent determines whether an executor is a command executor by
comparing the executor's path with its own --launcher_dir.  Because
the --launcher_dir changes (by a SHA in the path) during every upgrade,
all command executors will be considered "custom" executors after an
upgrade.  The agent will report executor resources to the master for
non-command executors only.