Add additional check inside the upgrade test.
This bit of logic grabs the state (from /master/state) before
and after the upgrade.  It then expects the two dictionaries to
be identical.  The master gets this information from the agent,
and as long as the agent does not change the information (which
it should not do between upgrades), the state should be identical.